src/requirements.txt
Django>=4.2,<4.3
gunicorn
python-decouple
dj-database-url
psycopg2-binary